Move species enumeration (non-strict) into the base
